+++ tests/FSharp.Data.GraphQL.Tests/ExecutionTests.fs
@@ -3,7 +3,6 @@
 
 module FSharp.Data.GraphQL.Tests.ExecutionTests
 
-open System
 open Xunit
 open FSharp.Data.GraphQL
 open FSharp.Data.GraphQL.Types