+++ RELEASE_NOTES.md
@@ -66,3 +66,9 @@
 * **Breaking Change** The `ISubscriptionProvider` interface now has additional methods to publish to subscriptions by tag.
 * **Breaking Change** Renamed interface methods in `ISubscriptionProvider` from XxAsync to AsyncXx to reserve the former for TPL extensions.
 * Fixes for stream and defer execution.
+
+### 0.0.17-beta - January 11 2019
+* Execute deferred results in parallel
+
+### 0.0.18-beta - January 16 2019
+* Fix dispose of deferred/stream responses